backend/movieapi_app/serializers.py
from rest_framework import serializers # import serializers from DRF
from .models import Movie


class MovieSerializer(serializers.ModelSerializer):
    class Meta:
        model = Movie
        fields = '__all__'
        