backend/requirements.txt
asgiref==3.5.0
blinker==1.6.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==3.2.12
django-admin==2.0.2
django-cors-headers==4.3.0
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.14.0
excel-base==1.0.4
exceptiongroup==1.1.3
Flask==3.0.0
gyp==0.1
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
iniconfig==2.0.0
isoweek==1.3.3
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.3
more-itertools==8.10.0
netifaces==0.11.0
npx==0.1.1
numpy==1.26.1
oauthlib==3.2.0
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.9
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pytest==7.3.1
pytest-mock==3.10.0
pytest-print==0.3.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==5.4.1
requests==2.31.0
requests-oauthlib==1.3.1
screen==1.0.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.2
systemd-python==234
TimeConvert==3.0.13
tomli==2.0.1
tzlocal==5.2
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.1.0
wadllib==1.3.6
Werkzeug==3.0.1
xlwt==1.3.0
zipp==1.0.0
